Change ThreadChecker to SequencedTaskChecker in internal::Call
In preparation of running DirectTransport on a TaskQueue in unit-tests, change the thread-checkers to sequence-checkers. This is necessary for Mac and iOS, where the TaskQueue is guaranteed to run sequentially, but not guaranteed to do so on only one thread.

TODO: Add the relevant BUGs.

BUG=None

Review-Url: https://codereview.webrtc.org/2999973002
Cr-Original-Commit-Position: refs/heads/master@{#19404}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: f3f5c0e0317094a0c43749e9ab12cb8c5b181b94