Change ThreadChecker to SequencedTaskChecker in VideoReceiveStream
In preparation of running DirectTransport on a TaskQueue in unit-tests, change the thread-checkers to sequence-checkers. This is necessary for Mac and iOS, where the TaskQueue is guaranteed to run sequentially, but not guaranteed to do so on only one thread.

TODO: Add the relevant BUGs.

BUG=None

Review-Url: https://codereview.webrtc.org/2997853002
Cr-Original-Commit-Position: refs/heads/master@{#19408}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: a28122f5cdd4c21c5674297e2cce14c8da15f9bc