Fix a bug in the SDP parser in AppRTCMobile.
It would leave a trailing carriage return character in the payload
list, causing the preferred codec to appear twice if it was at the
back of the list originally. This causes problems down the line and
results in that codec not being negotiated successfully.

BUG=webrtc:8129

Review-Url: https://codereview.webrtc.org/3001363002
Cr-Original-Commit-Position: refs/heads/master@{#19552}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 6644c8b7337abcf4bb91d8658a056c9bce0f8dfe